Fix bug in LineRecordReader
The previous implementation of LineRecordReader did not work for the
multiple-file case: hasNext() did not check for a next file and
initialise a new iterator.

This implementation fixes the bug, and cleans up the code with a bulk
of the work done in a new method called readLine(). The unit test
LineReaderTest is also modified accordingly.